refactor : remove index.ts from types directory.
