+++ registry/src/generic-form/FormFieldWrapper.tsx
@@ -8,7 +8,7 @@ import {
 } from '@/components/ui/form';
 import { cn } from '@/lib/utils';
 import { FieldValues } from 'react-hook-form';
-import { FormFieldWrapperProps } from '@/types';
+import { FormFieldWrapperProps } from '@/types/FormFieldWrapper.types';
 import { JSX } from 'react';
 
 function FormFieldWrapper<T extends FieldValues>({
