+++ registry/src/test/GenericForm.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, fireEvent } from '@testing-library/react';
 import { describe, it, expect, vi } from 'vitest';
 import { z } from 'zod';
 import GenericForm from '../generic-form/GenericForm';
-import { FormFieldConfig } from '../types';
+import { FormFieldConfig } from '../types/FormField.types';
 
 describe('GenericForm', () => {
   const validationSchema = z.object({
