--- registry/src/types/index.ts
@@ -1,4 +1,0 @@
-export type { GenericFormProps } from './GenericForm.types';
-export type { FormActionsProps } from './FormActions.types';
-export type { FormFieldWrapperProps } from './FormFieldWrapper.types';
-export type { FormFieldConfig } from './FormField.types';
